Update README.md
# Microsoft-Project
Tic-tac-toe (American English), noughts and crosses (British English), or Xs and Os is a paper-and-pencil game for two players, X and O, who take turns marking the spaces in a 3×3 grid. The player who succeeds in placing three of their marks in a horizontal, vertical, or diagonal row is the winner.
All of us know this game very well as we all played it in our childhood. 
But after joining Mars Colonisation Program we (a team of 4 members from IIT Roorkee) were motivated to slighty modify this game.So we decided to add somemore features like :
a) Including a score board
b) Including an algorithm which can provide hint to current player for the next move.

Mars Colonisation Project 
 https://cherryjain03.github.io/MINCS-Microsoft-Project/

Team Members:
1) Cherry Jain
2) Saakshi Choudhary
3) Ishita Kaur
4)Neha Bobde